Cleaning up, fixing missed errors
There were still a few cases of the following:

* time variables not converted from int to double
* parentheses not removed from getter names
* Preloader != PreLoader

+++ lib/Game/RenderObject/GameObject.dart
@@ -10,7 +10,7 @@ class GameObject extends RenderObject
 {
   int frame = 0;
   int frameY = 0;
-  double lastframeupdate = 0;
+  double lastframeupdate = 0.0;
   Vector vector;
   bool onPlatform = false;
 
@@ -125,13 +125,13 @@ class GameObject extends RenderObject
   }
 
 
-  bool changeImage(AnimationFrames changeTo, int looptime, int framerate)
+  bool changeImage(AnimationFrames changeTo, double looptime, int framerate)
   {
     int oldframe = frame;
     // next frame?
     if (currentAnimation == changeTo)
     {
-      int diff = looptime - lastframeupdate;
+      double diff = looptime - lastframeupdate;
       //per 20 frames add 1
       int updateframe = (1000 / framerate).toInt();
       if(diff < updateframe)
